# src/main/webapp/docs/swagger.yaml
swagger: "2.0"
info:
  description: "An anime api where you can get anime information, photos, videos, news and merchandising. Also you can update your MyAnimeList account."
  version: 1.0.0
  title: "M.A.R.C"
  # put the contact info for your development or API team
  contact:
    email: "grupoproyectomarc@gmail.com"

  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"

# all paths will be relative to the basePath specified
# basePath: /v1
# Added by API Auto Mocking Plugin
# host: proyectomarc-161510.appspot.com
# basePath: /api
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /GrupoMARC/M.A.R.C/1.0.0
...
schemes:
  - "https"
paths:
...
  "/videos":
    get:
      tags:
      - "Recurso Video"
      summary: "Método que devuelve todos los vídeos"
      operationId: "getAllVideos"
      produces:
        - "application/json"
      parameters: []
      responses:
        200:
          description: "Listado de todos los vídeos"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Video"
        404:
          description: "Ningún resultado"
  "/videos/idAnime":
    get:
      tags:
      - "Recurso Video"
      summary: "Método que devuelve todos los vídeos de la id de un determinado anime"
      operationId: "getAllVideosById"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "id"
          description: "ID del anime"
          required: true
          type: "string"
      responses:
        200:
          description: "Listado de todos los vídeos de un determinado anime"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Video"
        404:
          description: "Ningún resultado"
  "/videos/titleAnime":
    get:
      tags:
      - "Recurso Video"
      summary: "Método que devuelve todos los vídeos del nombre de un determinado anime"
      operationId: "getAllVideosByTitle"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "title"
          description: "Título del anime"
          required: true
          type: "string"
      responses:
        200:
          description: "Listado de todos los vídeos de un determinado anime"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Video"
        404:
          description: "Ningún resultado"
  "/videos/{id}":
    get:
      tags:
      - "Recurso Video"
      summary: "Método que devuelve un vídeo con un determinado id"
      operationId: "getVideoById"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "ID del vídeo"
          required: true
          type: "string"
      responses:
        200:
          description: "Vídeo según un determinado id"
          schema:
            $ref: "#/definitions/Video"
        404:
          description: "Ningún resultado"
  "/productos":
    get:
      tags:
      - "Recurso Producto"
      summary: "Método que devuelve todos los productos"
      operationId: "getAllProductos"
      produces:
        - "application/json"
      parameters: []
      responses:
        200:
          description: "Listado de todos los productos"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Producto"
        404:
          description: "Ningún resultado"
  "/productos/idAnime":
    get:
      tags:
      - "Recurso Producto"
      summary: "Método que devuelve todos los productos de la id de un determinado anime"
      operationId: "getAllProductosById"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "id"
          description: "ID del anime"
          required: true
          type: "string"
      responses:
        200:
          description: "Listado de todos los productos de un determinado anime"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Producto"
        404:
          description: "Ningún resultado"
  "/productos/titleAnime":
    get:
      tags:
      - "Recurso Producto"
      summary: "Método que devuelve todos los productos del nombre de un determinado anime"
      operationId: "getAllProductosByTitle"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "title"
          description: "Título del anime"
          required: true
          type: "string"
      responses:
        200:
          description: "Listado de todos los productos de un determinado anime"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Producto"
        404:
          description: "Ningún resultado"
  "/productos/{id}":
    get:
      tags:
      - "Recurso Producto"
      summary: "Método que devuelve un vídeo con un determinado id"
      operationId: "getProductoById"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "ID del vídeo"
          required: true
          type: "string"
      responses:
        200:
          description: "Vídeo según un determinado id"
          schema:
            $ref: "#/definitions/Producto"
        404:
          description: "Ningún resultado" 
...
definitions:
...
  Producto:
    required:
      - id
      - name
      - url
      - prize
      - description
    properties:
      id:
        type: string
        example: p1
      name:
        type: string
        example:  Cosplay Rika
      url:
        type: string
        example: https://es.aliexpress.com/item/Higurashi-no-Naku-Koro-ni-Higurashi-When-They-Cry-Rika-Furude-red-Witch-kimono-Cosplay-Costume/32350149865.html?spm=2114.43010208.4.1.P4V9SM
      prize:
        type: string
        example: 57,68 € 
      description:
        type: string
        example: Higurashi no Naku Koro ni Higurashi When They Cry Rika Furude red Witch kimono Cosplay Costume anysize
  Video:
    required:
      - id
      - title
      - url
    properties:
      id:
        type: string
        example: v1
      title:
        type: string
        example:  Higurashi no Naku Koro ni by Eiko Shimamiya
      url:
        type: string
        example: https://www.youtube.com/watch?v=XgWkIrFQQE0
...